+++ src/packetforge.cpp
@@ -22,6 +22,7 @@ PacketForge::PacketForge()
     {
         decryptPkt = reinterpret_cast<Decrypt>(
                             pktDecrypt.resolve( "decryptPacket" ) );
+
         initialized = true;
         message =  "Initialized PacketForge module with method "
                    "[ decryptPacket ].";
@@ -49,7 +50,7 @@ PacketForge* PacketForge::getInstance()
     return instance;
 }
 
-QString PacketForge::decryptPacket(const QString& packet)
+QString PacketForge::decryptPacket(const QByteArray& packet)
 {
     //Player positioning packets, return an empty string.
     if ( !Helper::strStartsWithStr( packet, ":SR?" )
@@ -63,7 +64,7 @@ QString PacketForge::decryptPacket(const QString& packet)
     return QString( "" );
 }
 
-bool PacketForge::validateSerNum(Player* plr, const QString& packet)
+bool PacketForge::validateSerNum(Player* plr, const QByteArray& packet)
 {
     QString pkt{ this->decryptPacket( packet ) };
 
@@ -73,6 +74,7 @@ bool PacketForge::validateSerNum(Player* plr, const QString& packet)
         return true;
 
     QString srcSerNum = pkt.left( 12 ).mid( 4 );
+
     //Unable to extract a SerNum from the incoming packet,
     //mark as valid.
     if ( srcSerNum.isEmpty() )