+++ src/widgets/chatviewwidget.hpp
@@ -34,7 +34,7 @@ class ChatViewWidget : public QWidget
         void setGameID(const Games& gID);
         Games getGameID() const;
 
-        void parsePacket(const QString& packet, const QString& alias = "");
+        void parsePacket(const QByteArray& packet, const QString& alias = "");
         void parseChatEffect(const QString& packet);
         void insertChat(const QString& msg, const Colors& color,
                             const bool& newLine);