+++ ChangeLog.md
@@ -6,10 +6,16 @@ Version 2.4.5:
     Change:
       * Moved Server specific information from the main ReMix UI to a ToolTip that is shown when hovering over a Server tab instance.
       * Added various log messages to further inform the Server Host as to what is happening in the background.
+      * The ChatView will now attempt to use the ingame name of a Player as opposed to just their sernum or alias.
 
 
     Bugfixes:
       * The Server Host will now be informed when the PacketForge Module is or is not able to be loaded.
+      * Packets sent to a Server Instance using the game type "Arcadia" will now properly decrypt packets when the PacketForge module is loaded.
+      * ReMix will now properly select the correct MasterMix server to correspond to the GameType selected via the CreateInstance dialog.
+      * ReMix will no longer attempt to continuously check-in with the "Warpath[W97]" MasterMix server.
+        * The "Warpath[W97]" MasterMix does not support responding to the Server Check-In packets.
+          * This is due to the MasterMix version being much older and not really compatible.